#rb none
#rnx

#ROBOMERGE-OWNER: patrick.boutot
#ROBOMERGE-AUTHOR: jeanmichel.dignard
#ROBOMERGE-SOURCE: CL 11099277 in //UE4/Dev-Tools-Staging/...
#ROBOMERGE-BOT: TOOLS (Dev-Tools-Staging -> Dev-VirtualProduction) (v640-11091645)

[CL 11114275 by patrick boutot in Dev-VirtualProduction branch]

+++ Engine/Source/Editor/TextureEditor/Private/Models/TextureEditorViewportClient.cpp
@@ -205,11 +205,11 @@ void FTextureEditorViewportClient::Draw(FViewport* Viewport, FCanvas* Canvas)
 	{
 		const uint32 Mip = (uint32)MipLevel;
 		const FIntPoint SizeOnMip = { Texture2D->GetSizeX() >> Mip,Texture2D->GetSizeY() >> Mip };
-		const uint64 NumPixels = SizeOnMip.X * SizeOnMip.Y;
+		const uint64 NumPixels = static_cast<uint64>(SizeOnMip.X) * SizeOnMip.Y;
 
 		const FVirtualTexture2DResource* Resource = (FVirtualTexture2DResource*)Texture2D->Resource;
 		const FIntPoint PhysicalTextureSize = Resource->GetPhysicalTextureSize(0u);
-		const uint64 NumPhysicalPixels = PhysicalTextureSize.X * PhysicalTextureSize.Y;
+		const uint64 NumPhysicalPixels = static_cast<uint64>(PhysicalTextureSize.X) * PhysicalTextureSize.Y;
 
 		if (NumPixels >= NumPhysicalPixels)
 		{
